Move manager interfaces, payload into domain package
