+++ app/src/main/java/au/com/safetychampion/Application.kt
@@ -1,7 +1,7 @@
 package au.com.safetychampion
 
 import au.com.safetychampion.data.di.dataModule
-import au.com.safetychampion.data.util.IDispatchers
+import au.com.safetychampion.data.domain.manager.IDispatchers
 import au.com.safetychampion.di.appModules
 import au.com.safetychampion.di.commonModule
 import au.com.safetychampion.util.koinGet
