+++ app/src/main/java/au/com/safetychampion/di/CommonModule.kt
@@ -1,6 +1,6 @@
 package au.com.safetychampion.di
 
-import au.com.safetychampion.data.util.* // ktlint-disable no-wildcard-imports
+import au.com.safetychampion.data.domain.manager.* // ktlint-disable no-wildcard-imports
 import au.com.safetychampion.local.AppDataStore
 import au.com.safetychampion.util.* // ktlint-disable no-wildcard-imports
 import kotlinx.coroutines.CoroutineDispatcher
