+++ app/src/main/java/au/com/safetychampion/util/TokenManager.kt
@@ -1,7 +1,7 @@
 package au.com.safetychampion.util
 
+import au.com.safetychampion.data.domain.manager.ITokenManager
 import au.com.safetychampion.data.domain.uncategory.AppToken
-import au.com.safetychampion.data.util.ITokenManager
 import au.com.safetychampion.dispatchers
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
