+++ data/src/main/java/au/com/safetychampion/data/data/common/TaskRepositoryImpl.kt
@@ -3,10 +3,10 @@ package au.com.safetychampion.data.data.common
 import au.com.safetychampion.data.data.BaseRepository
 import au.com.safetychampion.data.domain.core.Result
 import au.com.safetychampion.data.domain.models.TaskAssignStatusItem
+import au.com.safetychampion.data.domain.models.action.ActiveTaskPL
+import au.com.safetychampion.data.domain.models.action.AssignTaskStatusManyPL
+import au.com.safetychampion.data.domain.models.action.AssignTaskStatusPL
 import au.com.safetychampion.data.domain.models.task.Task
-import au.com.safetychampion.data.domain.payload.ActiveTaskPL
-import au.com.safetychampion.data.domain.payload.AssignTaskStatusManyPL
-import au.com.safetychampion.data.domain.payload.AssignTaskStatusPL
 
 class TaskRepositoryImpl(
     private val api: TaskAPI
