+++ data/src/main/java/au/com/safetychampion/data/domain/Attachment.kt
@@ -2,7 +2,7 @@ package au.com.safetychampion.data.domain
 
 import android.content.Intent
 import android.net.Uri
-import au.com.safetychampion.data.util.IFileManager
+import au.com.safetychampion.data.domain.manager.IFileManager
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import okhttp3.MultipartBody
