+++ data/src/main/java/au/com/safetychampion/data/domain/manager/IGsonManager.kt
@@ -1,4 +1,4 @@
-package au.com.safetychampion.data.util
+package au.com.safetychampion.data.domain.manager
 
 import com.google.gson.Gson
 import com.google.gson.GsonBuilder
