+++ data/src/main/java/au/com/safetychampion/data/domain/manager/ITokenManager.kt
@@ -1,4 +1,4 @@
-package au.com.safetychampion.data.util
+package au.com.safetychampion.data.domain.manager
 
 import au.com.safetychampion.data.domain.uncategory.AppToken
