+++ data/src/main/java/au/com/safetychampion/data/domain/models/action/ActionSignOff.kt
@@ -1,7 +1,5 @@
 package au.com.safetychampion.data.domain.models.action
 
-import au.com.safetychampion.data.domain.payload.ActionPojo
-
 data class ActionSignOff(
     val body: ActionPojo?,
     val _id: String?,
