+++ data/src/main/java/au/com/safetychampion/data/domain/models/action/AssignTaskStatusPL.kt
@@ -1,4 +1,4 @@
-package au.com.safetychampion.data.domain.payload
+package au.com.safetychampion.data.domain.models.action
 
 import au.com.safetychampion.data.domain.base.BasePL
 import au.com.safetychampion.data.domain.models.CreatedBy
