+++ data/src/main/java/au/com/safetychampion/data/domain/models/reviewplan/ReviewPlanSignoffTask.kt
@@ -5,8 +5,8 @@ import androidx.databinding.ObservableField
 import androidx.lifecycle.MutableLiveData
 import au.com.safetychampion.data.domain.InspectionFormPayload
 import au.com.safetychampion.data.domain.models.action.ActionLink
+import au.com.safetychampion.data.domain.models.action.ActionPojo
 import au.com.safetychampion.data.domain.models.customvalues.CustomValue
-import au.com.safetychampion.data.domain.payload.ActionPojo
 import au.com.safetychampion.data.domain.uncategory.Constants
 import au.com.safetychampion.data.domain.uncategory.DocAttachment
 import com.google.gson.annotations.SerializedName
