+++ data/src/main/java/au/com/safetychampion/data/domain/models/safetyplan/SafetyPlanSignoffTask.kt
@@ -5,7 +5,7 @@ import androidx.databinding.ObservableLong
 import androidx.lifecycle.MutableLiveData
 import au.com.safetychampion.data.domain.models.Tier
 import au.com.safetychampion.data.domain.models.action.ActionLink
-import au.com.safetychampion.data.domain.payload.ActionPojo
+import au.com.safetychampion.data.domain.models.action.ActionPojo
 import au.com.safetychampion.data.domain.uncategory.Constants
 import au.com.safetychampion.data.domain.uncategory.DocAttachment
 import com.google.gson.annotations.SerializedName
