+++ data/src/main/java/au/com/safetychampion/data/domain/usecase/action/GetListActionUseCase.kt
@@ -3,7 +3,7 @@ package au.com.safetychampion.data.domain.usecase.action
 import au.com.safetychampion.data.data.action.ActionRepository
 import au.com.safetychampion.data.domain.base.BasePL
 import au.com.safetychampion.data.domain.core.Result
-import au.com.safetychampion.data.domain.payload.ActionPojo
+import au.com.safetychampion.data.domain.models.action.ActionPojo
 
 class GetListActionUseCase(
     private val repository: ActionRepository
