+++ data/src/main/java/au/com/safetychampion/data/domain/usecase/action/SignOffActionUseCase.kt
@@ -2,8 +2,8 @@ package au.com.safetychampion.data.domain.usecase.action
 
 import au.com.safetychampion.data.data.action.ActionRepository
 import au.com.safetychampion.data.domain.core.Result
+import au.com.safetychampion.data.domain.models.action.ActionPojo
 import au.com.safetychampion.data.domain.models.action.ActionSignOffPL
-import au.com.safetychampion.data.domain.payload.ActionPojo
 
 class SignOffActionUseCase(
     private val repository: ActionRepository
