+++ data/src/main/java/au/com/safetychampion/data/domain/usecase/activetask/GetAllActiveTaskUseCase.kt
@@ -2,8 +2,8 @@ package au.com.safetychampion.data.domain.usecase.activetask
 
 import au.com.safetychampion.data.data.common.TaskRepository
 import au.com.safetychampion.data.domain.core.Result
+import au.com.safetychampion.data.domain.models.action.ActiveTaskPL
 import au.com.safetychampion.data.domain.models.task.Task
-import au.com.safetychampion.data.domain.payload.ActiveTaskPL
 
 class GetAllActiveTaskUseCase(
     private val activeTaskRepository: TaskRepository
