+++ data/src/main/java/au/com/safetychampion/data/domain/usecase/activetask/UnAssignTaskUseCase.kt
@@ -2,8 +2,8 @@ package au.com.safetychampion.data.domain.usecase.activetask
 
 import au.com.safetychampion.data.data.common.TaskRepository
 import au.com.safetychampion.data.domain.core.Result
+import au.com.safetychampion.data.domain.models.action.AssignTaskStatusPL
 import au.com.safetychampion.data.domain.models.task.Task
-import au.com.safetychampion.data.domain.payload.AssignTaskStatusPL
 
 class UnAssignTaskUseCase(
     private val activeTaskRepository: TaskRepository
