+++ data/src/main/java/au/com/safetychampion/data/domain/usecase/assigntaskstatus/AssignTaskStatusItemUseCase.kt
@@ -3,8 +3,8 @@ package au.com.safetychampion.data.domain.usecase.assigntaskstatus
 import au.com.safetychampion.data.data.common.TaskRepository
 import au.com.safetychampion.data.domain.core.Result
 import au.com.safetychampion.data.domain.models.TaskAssignStatusItem
+import au.com.safetychampion.data.domain.models.action.AssignTaskStatusPL
 import au.com.safetychampion.data.domain.models.task.Task
-import au.com.safetychampion.data.domain.payload.AssignTaskStatusPL
 
 class AssignTaskStatusItemUseCase(
     private val activeTaskRepository: TaskRepository
